refactor(cli/server): extract serialization helpers into Serializer module
Server was approaching CodeScene code health thresholds: 1265 lines,
75+ functions. Extracting the serialization layer into a dedicated
Serializer module keeps both modules well under the limits and separates
protocol-dispatch concerns from data-shaping concerns.

All public entry points (serialize_character, serialize_building_choices,
serialize_choices_list, etc.) are unchanged in behaviour; callers updated
to use Serializer.* prefix. resolve_character stays private in Server as
it is used by multiple handlers and does not belong in the serialization
layer.

refactor(cli/server): extract serialization helpers into Serializer module

Server was approaching CodeScene code health thresholds: 1265 lines,
75+ functions. Extracting the serialization layer into a dedicated
Serializer module keeps both modules well under the limits and separates
protocol-dispatch concerns from data-shaping concerns.

All public entry points (serialize_character, serialize_building_choices,
serialize_choices_list, etc.) are unchanged in behaviour; callers updated
to use Serializer.* prefix. resolve_character stays private in Server as
it is used by multiple handlers and does not belong in the serialization
layer.
