+++ apps/ttrpg_dev_cli/lib/cli/server.ex
@@ -43,8 +43,8 @@ defmodule ExTTRPGDev.CLI.Server do
   alias DiceLib.Basic, as: Dice
   alias ExTTRPGDev.Characters
   alias ExTTRPGDev.Characters.{Character, InventoryItem}
-  alias ExTTRPGDev.CLI.ConceptDisplay
-  alias ExTTRPGDev.RuleSystem.{Evaluator, InventoryRules}
+  alias ExTTRPGDev.CLI.Serializer
+  alias ExTTRPGDev.RuleSystem.Evaluator
   alias ExTTRPGDev.RuleSystems
   alias ExTTRPGDev.RuleSystems.LoadedSystem
 
@@ -128,10 +128,10 @@ defmodule ExTTRPGDev.CLI.Server do
             %{id: concept_id, concept_type: concept_type, fields: meta}
 
           concept_type != nil ->
-            serialize_concepts(system, concept_type)
+            Serializer.serialize_concepts(system, concept_type)
 
           true ->
-            serialize_system(system)
+            Serializer.serialize_system(system)
         end
 
       {ok(data), state}
@@ -162,7 +162,7 @@ defmodule ExTTRPGDev.CLI.Server do
 
       data =
         Map.put(
-          serialize_character(system, character, nil, parse_display_mode(msg)),
+          Serializer.serialize_character(system, character, nil, parse_display_mode(msg)),
           :temp_id,
           temp_id
         )
@@ -189,7 +189,7 @@ defmodule ExTTRPGDev.CLI.Server do
       pending = Map.put(state.pending, temp_id, character)
       new_state = %{state | pending: pending, next_id: state.next_id + 1}
 
-      building_choices = serialize_building_choices(system)
+      building_choices = Serializer.serialize_building_choices(system)
       {ok(%{temp_id: temp_id, building_choices: building_choices}), new_state}
     rescue
       e -> {error(Exception.message(e)), state}
@@ -212,7 +212,12 @@ defmodule ExTTRPGDev.CLI.Server do
       updated = %{character | decisions: character.decisions ++ [decision]}
 
       sub_choices =
-        serialize_concept_sub_choices(concept_type, concept_id, updated.decisions, system)
+        Serializer.serialize_concept_sub_choices(
+          concept_type,
+          concept_id,
+          updated.decisions,
+          system
+        )
 
       new_state = %{state | pending: Map.put(state.pending, temp_id, updated)}
       {ok(%{sub_choices: sub_choices}), new_state}
@@ -236,12 +241,15 @@ defmodule ExTTRPGDev.CLI.Server do
       character = fetch_pending!(state, temp_id)
       system = RuleSystems.load_system!(character.metadata.rule_system)
       scope = {scope_type, scope_id}
-      choice_def = fetch_choice_def!(system, scope, choice_id)
-      valid = valid_sub_choices(system, scope, choice_def, character.decisions)
+      choice_def = Serializer.fetch_choice_def!(system, scope, choice_id)
+      valid = Serializer.valid_sub_choices(system, scope, choice_def, character.decisions)
       validate_concept_selection!(selection, valid)
       decision = %{scope: scope, choice: choice_id, selection: selection}
       updated = %{character | decisions: character.decisions ++ [decision]}
-      sub_choices = serialize_concept_sub_choices(scope_type, scope_id, updated.decisions, system)
+
+      sub_choices =
+        Serializer.serialize_concept_sub_choices(scope_type, scope_id, updated.decisions, system)
+
       new_state = %{state | pending: Map.put(state.pending, temp_id, updated)}
       {ok(%{sub_choices: sub_choices}), new_state}
     rescue
@@ -261,8 +269,8 @@ defmodule ExTTRPGDev.CLI.Server do
       resolved = resolve_character(sys, updated)
       choices = Characters.pending_choices(sys, updated, resolved)
       mode = parse_display_mode(msg)
-      ser = serialize_character(sys, updated, updated.metadata.slug, mode)
-      data = Map.put(ser, :pending_choices, serialize_choices_list(choices, sys, mode))
+      ser = Serializer.serialize_character(sys, updated, updated.metadata.slug, mode)
+      data = Map.put(ser, :pending_choices, Serializer.serialize_choices_list(choices, sys, mode))
 
       {ok(data), new_state}
     rescue
@@ -336,10 +344,10 @@ defmodule ExTTRPGDev.CLI.Server do
       choices = Characters.pending_choices(system, updated, resolved)
 
       data =
-        serialize_character(system, updated, slug, parse_display_mode(msg))
+        Serializer.serialize_character(system, updated, slug, parse_display_mode(msg))
         |> Map.put(
           :pending_choices,
-          serialize_choices_list(choices, system, parse_display_mode(msg))
+          Serializer.serialize_choices_list(choices, system, parse_display_mode(msg))
         )
 
       {ok(data), state}
@@ -370,10 +378,10 @@ defmodule ExTTRPGDev.CLI.Server do
       choices = Characters.pending_choices(system, updated, resolved)
 
       data =
-        serialize_character(system, updated, slug, parse_display_mode(msg))
+        Serializer.serialize_character(system, updated, slug, parse_display_mode(msg))
         |> Map.put(
           :pending_choices,
-          serialize_choices_list(choices, system, parse_display_mode(msg))
+          Serializer.serialize_choices_list(choices, system, parse_display_mode(msg))
         )
         |> Map.put(:awarded_xp, xp_needed)
 
@@ -391,8 +399,10 @@ defmodule ExTTRPGDev.CLI.Server do
       resolved = resolve_character(system, character)
       choices = Characters.pending_choices(system, character, resolved)
 
-      {ok(%{pending_choices: serialize_choices_list(choices, system, parse_display_mode(msg))}),
-       state}
+      {ok(%{
+         pending_choices:
+           Serializer.serialize_choices_list(choices, system, parse_display_mode(msg))
+       }), state}
     rescue
       e -> {error(Exception.message(e)), state}
     end
@@ -468,10 +478,10 @@ defmodule ExTTRPGDev.CLI.Server do
       choices = Characters.pending_choices(system, updated, resolved)
 
       data =
-        serialize_character(system, updated, slug, parse_display_mode(msg))
+        Serializer.serialize_character(system, updated, slug, parse_display_mode(msg))
         |> Map.put(
           :pending_choices,
-          serialize_choices_list(choices, system, parse_display_mode(msg))
+          Serializer.serialize_choices_list(choices, system, parse_display_mode(msg))
         )
 
       {ok(data), state}
@@ -491,8 +501,8 @@ defmodule ExTTRPGDev.CLI.Server do
       Characters.save_character!(updated, true)
 
       data =
-        serialize_character(system, updated, slug, parse_display_mode(msg))
-        |> Map.put(:resolutions, serialize_resolutions(resolutions, system))
+        Serializer.serialize_character(system, updated, slug, parse_display_mode(msg))
+        |> Map.put(:resolutions, Serializer.serialize_resolutions(resolutions, system))
 
       {ok(data), state}
     rescue
@@ -511,7 +521,7 @@ defmodule ExTTRPGDev.CLI.Server do
     try do
       character = Characters.load_character!(slug)
       system = RuleSystems.load_system!(character.metadata.rule_system)
-      data = serialize_character(system, character, slug, parse_display_mode(msg))
+      data = Serializer.serialize_character(system, character, slug, parse_display_mode(msg))
       {ok(data), state}
     rescue
       e -> {error(Exception.message(e)), state}
@@ -555,7 +565,7 @@ defmodule ExTTRPGDev.CLI.Server do
   defp handle(%{"command" => "characters.inventory", "character" => slug}, state) do
     try do
       character = Characters.load_character!(slug)
-      {ok(%{inventory: serialize_inventory(character.inventory)}), state}
+      {ok(%{inventory: Serializer.serialize_inventory(character.inventory)}), state}
     rescue
       e -> {error(Exception.message(e)), state}
     end
@@ -580,7 +590,7 @@ defmodule ExTTRPGDev.CLI.Server do
         {:ok, item} ->
           updated = %{character | inventory: character.inventory ++ [item]}
           Characters.save_character!(updated, true)
-          {ok(%{inventory: serialize_inventory(updated.inventory)}), state}
+          {ok(%{inventory: Serializer.serialize_inventory(updated.inventory)}), state}
 
         {:error, reason} ->
           {error("cannot add item: #{inspect(reason)}"), state}
@@ -613,7 +623,7 @@ defmodule ExTTRPGDev.CLI.Server do
           new_inventory = List.replace_at(character.inventory, index, updated_item)
           updated = %{character | inventory: new_inventory}
           Characters.save_character!(updated, true)
-          {ok(%{inventory: serialize_inventory(updated.inventory)}), state}
+          {ok(%{inventory: Serializer.serialize_inventory(updated.inventory)}), state}
 
         {:error, reason} ->
           {error("cannot set field: #{inspect(reason)}"), state}
@@ -657,10 +667,10 @@ defmodule ExTTRPGDev.CLI.Server do
       choices = Characters.pending_choices(system, updated, resolved)
 
       data =
-        serialize_character(system, updated, slug, parse_display_mode(msg))
+        Serializer.serialize_character(system, updated, slug, parse_display_mode(msg))
         |> Map.put(
           :pending_choices,
-          serialize_choices_list(choices, system, parse_display_mode(msg))
+          Serializer.serialize_choices_list(choices, system, parse_display_mode(msg))
         )
 
       {ok(data), state}
@@ -676,286 +686,6 @@ defmodule ExTTRPGDev.CLI.Server do
   defp handle(_, state) do
     {error("request must have a \"command\" field"), state}
   end
-
-  # --- Serialization ---
-
-  defp serialize_system(%LoadedSystem{module: mod}) do
-    %{
-      name: mod.name,
-      slug: mod.slug,
-      version: mod.version,
-      publisher: mod.publisher,
-      family: mod.family,
-      series: mod.series,
-      concept_types: Enum.map(mod.concept_types, &%{id: &1.id, name: &1.name})
-    }
-  end
-
-  defp serialize_concepts(%LoadedSystem{concept_metadata: meta}, concept_type) do
-    concepts =
-      meta
-      |> Enum.filter(fn {{type, _id}, _} -> type == concept_type end)
-      |> Enum.sort_by(fn {{_type, id}, _} -> id end)
-      |> Enum.map(fn {{_type, id}, fields} ->
-        %{id: id, name: Map.get(fields, "name", id)}
-      end)
-
-    %{concept_type: concept_type, concepts: concepts}
-  end
-
-  defp serialize_character(%LoadedSystem{} = system, %Character{} = character, slug, display_mode) do
-    active = Characters.active_concepts(character.decisions, system.concept_metadata)
-    resolved = resolve_character(system, character)
-    resolved_by_concept = Enum.group_by(resolved, fn {{type, id, _field}, _} -> {type, id} end)
-    inventory_ids = MapSet.new(character.inventory, &{&1.concept_type, &1.concept_id})
-
-    %{
-      name: character.name,
-      rule_system: character.metadata.rule_system,
-      slug: slug,
-      choices: serialize_choices(system, character),
-      character_lists: serialize_character_lists(system, character, active, display_mode),
-      concept_types:
-        serialize_concept_type_values(system, resolved_by_concept, inventory_ids, active),
-      selected_concepts: serialize_selected_concepts(system, character, display_mode)
-    }
-  end
-
-  defp serialize_selected_concepts(
-         %LoadedSystem{} = system,
-         %Character{} = character,
-         display_mode
-       ) do
-    selection_progressions =
-      system.concept_metadata
-      |> Enum.filter(fn {{type, _id}, meta} ->
-        type == "character_progression" and Map.has_key?(meta, "type")
-      end)
-      |> Map.new(fn {{_type, id}, meta} ->
-        {id, %{concept_type: meta["type"], name: meta["name"] || id}}
-      end)
-
-    character.decisions
-    |> Enum.filter(fn
-      %{scope: {"character_progression", prog_id}} ->
-        Map.has_key?(selection_progressions, prog_id)
-
-      _ ->
-        false
-    end)
-    |> Enum.map(fn %{scope: {"character_progression", prog_id}, selection: selection} ->
-      prog = selection_progressions[prog_id]
-      {prog.concept_type, prog.name, selection}
-    end)
-    |> Enum.uniq()
-    |> Enum.map(fn {concept_type, progression_name, id} ->
-      meta = system.concept_metadata[{concept_type, id}] || %{"name" => id}
-      template = find_display_template(system, concept_type)
-      label = ConceptDisplay.render(template, meta, display_mode)
-      %{progression: progression_name, id: id, label: label}
-    end)
-    |> Enum.sort_by(fn %{progression: prog, label: label} -> {prog, label} end)
-  end
-
-  defp serialize_choices(%LoadedSystem{} = system, %Character{} = character) do
-    system.module.character_building_choices
-    |> Enum.flat_map(&serialize_choice_entry(system, character, &1))
-  end
-
-  defp serialize_choice_entry(system, character, %{concept_type: type_id}) do
-    root = Enum.find(character.decisions, &(&1.scope == nil and &1.choice == type_id))
-
-    if root do
-      type_name =
-        Enum.find_value(system.module.concept_types, &if(&1.id == type_id, do: &1.name))
-
-      chain =
-        concept_name_chain(character.decisions, system.concept_metadata, type_id, root.selection)
-
-      [%{type_name: type_name, value: Enum.join(chain, " / ")}]
-    else
-      []
-    end
-  end
-
-  defp concept_name_chain(decisions, concept_metadata, type_id, concept_id) do
-    name = get_in(concept_metadata, [{type_id, concept_id}, "name"]) || concept_id
-
-    sub_names =
-      concept_metadata
-      |> Map.get({type_id, concept_id}, %{})
-      |> Map.get("choices", %{})
-      |> Enum.flat_map(&same_type_sub_chain(decisions, concept_metadata, type_id, concept_id, &1))
-
-    [name | sub_names]
-  end
-
-  defp same_type_sub_chain(
-         decisions,
-         concept_metadata,
-         type_id,
-         concept_id,
-         {choice_id, choice_def}
-       ) do
-    if choice_def["type"] == type_id do
-      decision =
-        Enum.find(decisions, &(&1.scope == {type_id, concept_id} and &1.choice == choice_id))
-
-      if decision,
-        do: concept_name_chain(decisions, concept_metadata, type_id, decision.selection),
-        else: []
-    else
-      []
-    end
-  end
-
-  defp serialize_character_lists(system, character, active, display_mode) do
-    ctx = {system, display_mode}
-
-    system.module.character_lists
-    |> Enum.map(&serialize_character_list_category(&1, ctx, character, active))
-    |> Enum.reject(fn %{items: items} -> items == [] end)
-  end
-
-  defp serialize_character_list_category(cat, {system, display_mode}, character, active) do
-    template = find_display_template(system, cat.concept_type)
-    ids = collect_from_active(active, system.concept_metadata, cat.metadata_key)
-
-    items =
-      case cat.concept_type do
-        nil ->
-          ids
-
-        concept_type ->
-          Enum.map(ids, fn id ->
-            meta = system.concept_metadata[{concept_type, id}] || %{"name" => id}
-            ConceptDisplay.render(template, meta, display_mode)
-          end)
-      end
-
-    choice_template = find_display_template(system, cat.choice_concept_type)
-
-    choice_items =
-      case cat.choice_concept_type do
-        nil ->
-          []
-
-        concept_type ->
-          character.decisions
-          |> chosen_by_type(system.concept_metadata, concept_type)
-          |> Enum.map(fn id ->
-            meta = system.concept_metadata[{concept_type, id}] || %{"name" => id}
-            ConceptDisplay.render(choice_template, meta, display_mode)
-          end)
-      end
-
-    %{label: cat.label, items: (items ++ choice_items) |> Enum.uniq() |> Enum.sort()}
-  end
-
-  defp collect_from_active(active, concept_metadata, key) do
-    active
-    |> Enum.flat_map(fn {type, id} ->
-      Map.get(concept_metadata[{type, id}] || %{}, key, [])
-    end)
-    |> Enum.uniq()
-    |> Enum.sort()
-  end
-
-  defp chosen_by_type(decisions, concept_metadata, type) do
-    decisions
-    |> Enum.filter(fn
-      %{scope: {scope_type, scope_id}, choice: choice_id} ->
-        choice_def =
-          get_in(concept_metadata, [{scope_type, scope_id}, "choices", choice_id]) || %{}
-
-        choice_def["type"] == type and choice_def["grants_to"] != "inventory"
-
-      _ ->
-        false
-    end)
-    |> Enum.map(& &1.selection)
-  end
-
-  defp serialize_inventory(inventory) do
-    inventory
-    |> Enum.with_index()
-    |> Enum.map(fn {%InventoryItem{} = item, index} ->
-      %{
-        index: index,
-        concept_type: item.concept_type,
-        concept_id: item.concept_id,
-        fields: item.fields
-      }
-    end)
-  end
-
-  defp serialize_concept_type_values(
-         %LoadedSystem{} = system,
-         resolved_by_concept,
-         inventory_ids,
-         active
-       ) do
-    choice_types =
-      system.module.character_building_choices
-      |> Enum.map(& &1.concept_type)
-      |> MapSet.new()
-
-    ctx = %{
-      concept_metadata: system.concept_metadata,
-      inventory_rules: system.inventory_rules,
-      inventory_ids: inventory_ids,
-      signed: MapSet.new(system.module.display_config.signed_fields),
-      choice_types: choice_types,
-      active: active
-    }
-
-    Enum.flat_map(
-      system.module.concept_types,
-      &serialize_concept_type(&1, resolved_by_concept, ctx)
-    )
-  end
-
-  defp serialize_concept_type(concept_type, resolved_by_concept, ctx) do
-    inventoriable = InventoryRules.inventoriable?(ctx.inventory_rules, concept_type.id)
-    choice_driven = MapSet.member?(ctx.choice_types, concept_type.id)
-
-    concepts =
-      ctx.concept_metadata
-      |> Enum.filter(fn {{type, _id}, _} -> type == concept_type.id end)
-      |> Enum.sort_by(fn {{_type, id}, _} -> id end)
-      |> Enum.filter(fn {{type, id}, meta} ->
-        Map.has_key?(resolved_by_concept, {type, id}) and
-          not Map.get(meta, "hidden", false) and
-          (not inventoriable or MapSet.member?(ctx.inventory_ids, {type, id})) and
-          (not choice_driven or MapSet.member?(ctx.active, {type, id}))
-      end)
-
-    if concepts == [] do
-      []
-    else
-      entries = Enum.map(concepts, &serialize_concept_entry(&1, resolved_by_concept, ctx.signed))
-      [%{id: concept_type.id, name: concept_type.name, concepts: entries}]
-    end
-  end
-
-  defp serialize_concept_entry({{type, id}, meta}, resolved_by_concept, signed) do
-    name = meta["name"] || id
-
-    fields =
-      resolved_by_concept[{type, id}]
-      |> Enum.sort_by(fn {{_t, _i, field}, _} -> field end)
-      |> Enum.map(fn {{_t, _i, field}, value} ->
-        %{name: field, value: format_field_value(field, value, signed)}
-      end)
-
-    %{id: id, name: name, fields: fields}
-  end
-
-  defp format_field_value(field, value, signed) when is_integer(value) and value >= 0 do
-    if MapSet.member?(signed, field), do: "+#{value}", else: "#{value}"
-  end
-
-  defp format_field_value(_field, value, _signed), do: "#{value}"
 
   defp resolve_character(%LoadedSystem{} = system, %Character{} = character) do
     system
@@ -1034,175 +764,6 @@ defmodule ExTTRPGDev.CLI.Server do
     )
   end
 
-  defp serialize_choices_list(choices, system, display_mode) do
-    Enum.map(choices, fn
-      %{type: :pending, options: options} = c ->
-        concept_type = pending_choice_concept_type(c, system)
-        template = find_display_template(system, concept_type)
-
-        rendered_options =
-          Enum.map(options, fn id ->
-            fields = system.concept_metadata[{concept_type, id}] || %{"name" => id}
-            %{id: id, label: ConceptDisplay.render(template, fields, display_mode)}
-          end)
-
-        base = %{
-          type: "pending",
-          id: c.id,
-          name: c.name,
-          count: c.count,
-          roll: Map.get(c, :roll),
-          options: rendered_options,
-          earned_at_level: Map.get(c, :earned_at_level)
-        }
-
-        scope_extras = %{scope_type: Map.get(c, :scope_type), scope_id: Map.get(c, :scope_id)}
-        Map.merge(base, Map.reject(scope_extras, fn {_, v} -> is_nil(v) end))
-
-      %{type: :pending} = c ->
-        %{type: "pending", id: c.id, name: c.name, count: c.count, roll: Map.get(c, :roll)}
-
-      %{type: :available} = c ->
-        %{type: "available", id: c.id, name: c.name, roll: Map.get(c, :roll)}
-    end)
-  end
-
-  defp pending_choice_concept_type(
-         %{scope_type: scope_type, scope_id: scope_id, id: choice_id},
-         system
-       ) do
-    get_in(system.concept_metadata, [{scope_type, scope_id}, "choices", choice_id, "type"])
-  end
-
-  defp pending_choice_concept_type(%{id: id}, system) do
-    get_in(system.concept_metadata, [{"character_progression", id}, "type"])
-  end
-
-  defp serialize_resolutions(resolutions, system) do
-    Enum.map(resolutions, fn r ->
-      selection_name =
-        r.concept_type &&
-          r.selection_id &&
-          get_in(system.concept_metadata, [{r.concept_type, r.selection_id}, "name"])
-
-      %{
-        name: r.name,
-        selection_id: r.selection_id,
-        selection_name: selection_name,
-        rolled_value: r.rolled_value,
-        method: r.method,
-        earned_at_level: r.earned_at_level
-      }
-    end)
-  end
-
-  defp find_display_template(system, concept_type) do
-    Enum.find_value(system.module.concept_types, fn ct ->
-      if ct.id == concept_type, do: ct.display_template
-    end)
-  end
-
-  defp serialize_concept_sub_choices(concept_type, concept_id, decisions, system) do
-    scope_key = {concept_type, concept_id}
-    choices = get_in(system.concept_metadata, [scope_key, "choices"]) || %{}
-
-    already_chosen_by_type =
-      decisions
-      |> Enum.filter(&(&1.scope == scope_key))
-      |> Enum.group_by(&(choices[&1.choice] || %{})["type"], & &1.selection)
-
-    choices
-    |> Enum.flat_map(fn {choice_id, choice_def} ->
-      resolved = Enum.count(decisions, &(&1.scope == scope_key and &1.choice == choice_id))
-
-      pending_count = if choice_def["required"] == true, do: max(0, 1 - resolved), else: 0
-
-      render_pending_sub_choice_entry(
-        choice_id,
-        choice_def,
-        concept_type,
-        concept_id,
-        pending_count,
-        already_chosen_by_type,
-        system
-      )
-    end)
-    |> Enum.sort_by(& &1.id)
-  end
-
-  defp render_pending_sub_choice_entry(_, _, _, _, 0, _, _), do: []
-
-  defp render_pending_sub_choice_entry(
-         choice_id,
-         choice_def,
-         concept_type,
-         concept_id,
-         pending_count,
-         already_chosen_by_type,
-         system
-       ) do
-    choice_type = choice_def["type"]
-    raw_options = build_sub_choice_options(choice_def, choice_type, system)
-    excluded = MapSet.new(Map.get(already_chosen_by_type, choice_type, []))
-    filtered = Enum.reject(raw_options, &MapSet.member?(excluded, &1))
-    template = find_display_template(system, choice_type)
-
-    rendered =
-      Enum.map(filtered, fn id ->
-        fields = system.concept_metadata[{choice_type, id}] || %{"name" => id}
-        %{id: id, label: ConceptDisplay.render(template, fields, :default)}
-      end)
-
-    [
-      %{
-        type: "pending",
-        id: choice_id,
-        scope_type: concept_type,
-        scope_id: concept_id,
-        name: choice_def["name"] || choice_id,
-        count: pending_count,
-        options: rendered
-      }
-    ]
-  end
-
-  defp build_sub_choice_options(choice_def, choice_type, system) do
-    case choice_def["options"] do
-      options when is_list(options) ->
-        options
-
-      _ ->
-        system.concept_metadata
-        |> Enum.filter(fn {{t, _}, _} -> t == choice_type end)
-        |> Enum.map(fn {{_, id}, _} -> id end)
-        |> Enum.sort()
-    end
-  end
-
-  defp fetch_choice_def!(system, {type, id}, choice_id) do
-    get_in(system.concept_metadata, [{type, id}, "choices", choice_id]) ||
-      raise("unknown choice #{inspect(choice_id)} on #{type}(#{id})")
-  end
-
-  defp valid_sub_choices(system, {scope_type, scope_id} = scope, choice_def, decisions) do
-    choice_type = choice_def["type"]
-    raw_options = build_sub_choice_options(choice_def, choice_type, system)
-
-    already_chosen =
-      decisions
-      |> Enum.filter(fn
-        %{scope: ^scope, choice: choice} ->
-          cd = get_in(system.concept_metadata, [{scope_type, scope_id}, "choices", choice]) || %{}
-          cd["type"] == choice_type
-
-        _ ->
-          false
-      end)
-      |> MapSet.new(& &1.selection)
-
-    Enum.reject(raw_options, &MapSet.member?(already_chosen, &1))
-  end
-
   defp fetch_pending!(state, temp_id) do
     Map.get(state.pending, temp_id) || raise("no pending character: #{inspect(temp_id)}")
   end
@@ -1212,29 +773,6 @@ defmodule ExTTRPGDev.CLI.Server do
     |> String.downcase()
     |> String.replace(~r/[!#$%&()*+,.:;<=>?@\^_`'{|}~-]/, "")
     |> String.replace(" ", "_")
-  end
-
-  defp serialize_building_choices(%LoadedSystem{} = system) do
-    Enum.map(system.module.character_building_choices, fn cc ->
-      concept_type = cc.concept_type
-      template = find_display_template(system, concept_type)
-
-      root_ids = MapSet.new(Characters.root_concept_ids(system.concept_metadata, concept_type))
-
-      concepts =
-        system.concept_metadata
-        |> Enum.filter(fn {{type, id}, _} ->
-          type == concept_type and MapSet.member?(root_ids, id)
-        end)
-        |> Enum.sort_by(fn {{_type, id}, _} -> id end)
-        |> Enum.map(fn {{_type, id}, fields} ->
-          %{id: id, label: ConceptDisplay.render(template, fields, :default)}
-        end)
-
-      types = system.module.concept_types
-      ct_name = Enum.find_value(types, concept_type, &(&1.id == concept_type && &1.name))
-      %{concept_type: concept_type, name: ct_name, concepts: concepts}
-    end)
   end
 
   defp parse_display_mode(msg) do
